refactor: split merge constraint detectors into FK/unique/CHECK
Split the ~1.8k-line doltlite_merge_constraints.c along the same seams
as the merge core: shared catalog/row helpers stay in
doltlite_merge_constraints.c with a private int header; unique, CHECK,
and FK detection each get their own module.

Wire main.mk, amalgamation, AGENTS.md, and layer lints so the split
must exist and each module stays under 1500 lines.

Verified: lint_layers, lint_layers_selftest; doltlite_merge and
doltlite_merge_status shell suites.
